@@ -11,7 +11,6 @@ import net.md_5.bungee.api.plugin.TabExecutor;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import quest.ender.Matchmaker.Matchmaker;
-import quest.ender.Matchmaker.util.PartyUtil;
 
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.TimeUnit;
@@ -39,37 +38,31 @@ public class MakeMatchCommand extends Command implements TabExecutor {
             final @NotNull ProxiedPlayer proxiedPlayer = (ProxiedPlayer) sender;
             final @NotNull String gameName = args[0];
 
-            if (!PartyUtil.leadsParty(proxiedPlayer)) {
-                sender.sendMessage(new ComponentBuilder("Only the leader of a party can queue you for a game!").color(ChatColor.RED).create());
+            final @Nullable CompletableFuture<ServerInfo> serverInfoCompletableFuture = this.matchmaker.sendToGame(proxiedPlayer, gameName);
+
+            if (serverInfoCompletableFuture != null) {
+                serverInfoCompletableFuture.thenApply((targetServerInfo) -> {
+                    sender.sendMessage(new ComponentBuilder("Sent you to " + targetServerInfo.getName() + ".").color(ChatColor.GREEN).create());
+                    return targetServerInfo;
+                });
+
+                this.matchmaker.getProxy().getScheduler().schedule(this.matchmaker, () -> {
+                    if (!serverInfoCompletableFuture.isDone()) {
+                        serverInfoCompletableFuture.cancel(true);
+                        sender.sendMessage(new ComponentBuilder("We were unable to send you to a game. Try again later.").color(ChatColor.RED).create());
+                    }
+                }, this.matchmaker.getConfig().getLong("timeout"), TimeUnit.MILLISECONDS);
             } else {
-                final @Nullable CompletableFuture<ServerInfo> serverInfoCompletableFuture = this.matchmaker.sendToGame(proxiedPlayer, gameName);
-
-                if (serverInfoCompletableFuture != null) {
-                    serverInfoCompletableFuture.thenApply((targetServerInfo) -> {
-                        sender.sendMessage(new ComponentBuilder("Sent you to " + targetServerInfo.getName() + ".").color(ChatColor.GREEN).create());
-                        return targetServerInfo;
-                    });
-
-                    this.matchmaker.getProxy().getScheduler().schedule(this.matchmaker, () -> {
-                        if (!serverInfoCompletableFuture.isDone()) {
-                            serverInfoCompletableFuture.cancel(true);
-                            sender.sendMessage(new ComponentBuilder("We were unable to send you to a game. Try again later.").color(ChatColor.RED).create());
-                        }
-                    }, this.matchmaker.getConfig().getLong("timeout"), TimeUnit.MILLISECONDS);
-                }
-
-                // Thread-safety is likely an issue here.
+                sender.sendMessage(new ComponentBuilder("Something prevented you from entering that game. You may not have permission to do so.").color(ChatColor.RED).create());
             }
         }
     }
 
     @Override
     public Iterable<String> onTabComplete(CommandSender sender, String[] args) {
-        switch (args.length) {
-            case 1:
-                return this.matchmaker.getGames();
-            default:
-                return null;
-        }
+        return switch (args.length) {
+            case 1 -> this.matchmaker.getGames();
+            default -> null;
+        };
     }
 }
